refactor AddEditProductDrawer
