+++ public/locales/en/translation.json
@@ -1,17 +1,24 @@
 {
   "All goods": "All goods",
   "Add category": "Add category",
-  "Add new item": "Add new item",
+  "Adding a new product": "Add new product",
+  "Add a link to the product image": "Add a link to the product image",
+  "Add image": "Add image",
   "Category has been edited": "Category has been edited",
   "Category list is empty": "Category list is empty",
   "Category has been removed": "Category has been removed",
   "Category has been successfully created": "Category has been successfully created",
+  "Description": "Description",
   "Free shipping": "Free shipping",
   "Favorites": "Favorites",
   "Failed to get list of categories": "Failed to get list of categories",
   "Failed to load category list": "Failed to load category list",
   "Failed to load product list": "Failed to load product list",
   "Failed to load product list. Please try again later.": "Failed to load product list. Please try again later.",
+  "Failed to load favorites list": "Failed to load favorites list",
+  "Failed to load list of favorite products. Please try again later.": "Failed to load list of favorite products. Please try again later.",
+  "Favorites list is empty": "Favorites list is empty",
+  "Go to catalog": "Go to catalog",
   "I already have an account": "I already have an account",
   "Language": "EN",
   "Log out": "Log out",
@@ -22,15 +29,23 @@
   "Profile": "Profile",
   "Price low to high": "Price low to high",
   "Price high to low": "Price high to low",
+  "Product category": "Product category",
+  "Product images": "Product images",
+  "Edit product": "Edit product",
+  "Save": "Save",
   "Search for items...": "Search for items...",
   "Service is temporarily unavailable": "Service is temporarily unavailable",
   "Settings": "Settings",
   "Sign in": "Sign in",
   "Sign up": "Sign up",
+  "Select a category": "Select a category",
+  "Saving...": "Saving...",
   "The items you order will be displayed here": "The items you order will be displayed here",
   "There are no products in this category": "There are no products in this category",
+  "There is nothing here yet, but you can add the product to your favorites by clicking on": "There is nothing here yet, but you can add the product to your favorites by clicking on",
   "User role has been successfully changed": "User role has been successfully changed",
   "View bag": "View bag",
+  "Vendor": "Vendor",
   "Wrong login or password": "Wrong login or password",
   "You have successfully logged in": "You have successfully logged in",
   "You have successfully signed up": "You have successfully signed up",
