+++ public/locales/ru/translation.json
@@ -2,16 +2,26 @@
   "Add new item": "Добавить новый товар",
   "Add category": "Добавить категорию",
   "All goods": "Все товары",
+  "Adding a new product": "Добавление нового товара",
+  "Add image": "Добавить изображение",
+  "Add a link to the product image": "Добавьте ссылку на изображение",
+  "Description": "Описание",
+  "Product category": "Категория товара",
   "Category has been successfully created": "Категория была успешно создана",
   "Category list is empty": "Список категорий пуст",
   "Category has been removed": "Категория была удалена",
   "Category has been edited": "Категория была отредактирована",
+  "Edit product": "Редактирование товара",
   "Free shipping": "Бесплатная доставка",
   "Failed to get list of categories": "Не удалось получить список категорий",
   "Failed to load category list": "Не удалось загрузить список категорий",
   "Failed to load product list": "Не удалось загрузить список товаров",
   "Failed to load product list. Please try again later.": "Не удалось загрузить список товаров. Повторите попытку позже.",
+  "Failed to load favorites list": "Не удалось загрузить список избранных товаров",
+  "Failed to load list of favorite products. Please try again later.": "Не удалось загрузить список избранных товаров. Повторите попытку позже.",
   "Favorites": "Избранное",
+  "Favorites list is empty": "В избранном ничего нет",
+  "Go to catalog": "В каталог",
   "I already have an account": "У меня уже есть аккаунт",
   "Language": "RU",
   "Log out": "Выйти",
@@ -22,14 +32,20 @@
   "Price low to high": "Сначала дешёвые",
   "Price high to low": "Сначала дорогие",
   "Profile": "Профиль",
-  "View bag": "Перейти в корзину",
+  "Product images": "Изображения товара",
   "Search for items...": "Поиск по товарам...",
   "Service is temporarily unavailable": "Сервис временно недоступен",
   "Sign in": "Войти",
   "Sign up": "Регистрация",
+  "Select a category": "Выберите категорию",
+  "Settings": "Настройки",
+  "Save": "Сохранить",
+  "Saving...": "Сохранение...",
+  "Vendor": "Продавец",
   "The items you order will be displayed here": "Тут появятся товары, которые вы закажете",
   "There are no products in this category": "В данной категории нет товаров",
-  "Settings": "Настройки",
+  "There is nothing here yet, but you can add the product to your favorites by clicking on": "Здесь пока ничего нет, но вы можете добавить товар в избранное, нажав на",
+  "View bag": "Перейти в корзину",
   "User role has been successfully changed": "Роль пользователя была успешно изменена",
   "Wrong login or password": "Неверный логин или пароль",
   "You have successfully signed up": "Вы успешно зарегистрировались",
